Fix RT#20388: Support whitespace after :include: in Sendmail format
As per https://rt.cpan.org/Ticket/Display.html?id=20388, sendmail
allows whitespace after `:include:` and before the filename in aliases
files, so we need to support that as well.

Note: That bug claims that before this fix the code was going into an
infinite loop, but I can't reproduce that. The broken behavior I was
seeing was that the alias was expanding to `:include: /filename`
instead of properly including the contents of `/filename`. In any
case, the include should work properly now.